+++ src/Modifier/QueryFilter.php
@@ -14,7 +14,6 @@ namespace League\Csv\Modifier;
 
 use ArrayObject;
 use CallbackFilterIterator;
-use InvalidArgumentException;
 use Iterator;
 use LimitIterator;
 
@@ -102,13 +101,15 @@ trait QueryFilter
      */
     public function setOffset($offset = 0)
     {
-        if (false === filter_var($offset, FILTER_VALIDATE_INT, ['options' => ['min_range' => 0]])) {
-            throw new InvalidArgumentException('the offset must be a positive integer or 0');
-        }
-        $this->iterator_offset = $offset;
+        $this->iterator_offset = $this->filterInteger($offset, 0, 'the offset must be a positive integer or 0');
 
         return $this;
     }
+
+    /**
+     * @inheritdoc
+     */
+    abstract protected function filterInteger($int, $min, $message);
 
     /**
      * Set LimitIterator Count
@@ -119,10 +120,7 @@ trait QueryFilter
      */
     public function setLimit($limit = -1)
     {
-        if (false === filter_var($limit, FILTER_VALIDATE_INT, ['options' => ['min_range' => -1]])) {
-            throw new InvalidArgumentException('the limit must an integer greater or equals to -1');
-        }
-        $this->iterator_limit = $limit;
+        $this->iterator_limit = $this->filterInteger($limit, -1, 'the limit must an integer greater or equals to -1');
 
         return $this;
     }